upim-core, edit: Collect, report all parse errors
I had to change the upim_core::error::FileError to store an
std::io::ErrorKind instead of the Error, since I need to be able to
copy/clone these and io::Error isn't cloneable.

This also fixes parsing an empty variable name in a config file, which didn't
return an error before.

The error-managing isn't bad but I can probably improve the API.